robot-classification/requirements.txt
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
filelock==3.25.2
filetype==1.2.0
fonttools==4.62.1
fsspec==2026.3.0
idna==3.7
ipykernel==7.2.0
ipython==8.39.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
opencv-python-headless==4.10.0.84

parso==0.8.6
pexpect==4.9.0
pi_heif==1.3.0
pillow==12.1.1
pillow-avif-plugin==1.5.5
platformdirs==4.9.4
polars==1.39.3
polars-runtime-32==1.39.3
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.0
requests-toolbelt==1.0.0
roboflow==1.2.16
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
ultralytics==8.4.31
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
